Group service properties that are temporally coherent.
Even though several properties is an attracting way to represent for example the x, y, and z axes of an accelerometer, it makes things more difficult for an observer that needs to be notified of these values at once.  After all, the "data" event happen once in the service.
Another way of modelling this would have been to keep the properties the way there were, but add a delegate method to advertise that an event was fired, but in my opinion that defeats the ability to have several observers.